# doc/api.yaml
openapi: 3.0.1
info:
  title: WASAText API
  version: 1.0.0
  description: API per l'app di messaggistica WASAText
servers:
  - url: http://0.0.0.0:3000
...
security:
  - bearerAuth: []
paths:
...
  /user/me:
    get:
      tags:
        - user
      summary: Ottieni le informazioni dell’utente autenticato
      description: Restituisce le generalità dell’utente attualmente autenticato, incluse UUID, username e URL foto profilo.
      operationId: getMyUserInfo
      responses:
        '200':
          description: Dati dell’utente autenticato restituiti con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/me/username:
    put:
      tags: 
        - user
      summary: Modifica lo username dell’utente autenticato
      description: Permette all’utente autenticato di modificare il proprio username.
      operationId: setMyUserName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 16
                  pattern: '^[a-zA-Z0-9_]+$'
                  description: Nuovo username
      responses:
        '200':
          description: Username aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/me/photo:
    put:
      tags: 
        - user
      summary: Modifica la foto profilo dell’utente autenticato
      description: Permette all’utente autenticato di aggiornare la propria foto profilo.
      operationId: setMyPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - photo
              properties:
                photo:
                  type: string
                  format: binary
                  description: URL della nuova foto profilo
      responses:
        '200':
          description: Foto profilo aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/all:
    get:
      tags:
        - user
      summary: Recupera tutti gli utenti registrati
      description: Restituisce una lista di tutti gli utenti attualmente registrati sulla piattaforma.
      operationId: getAllUsers
      responses:
        '200':
          description: Lista utenti restituita con successo
          content:
            application/json:
              schema:
                type: object
                required:
                  - users
                properties:
                  users:
                    type: array
                    minItems: 0
                    maxItems: 200
                    items:
                      type: object
                      properties:
                        username:
                          type: string
                          example: "luca_dev"
                        photoUrl:
                          type: string
                          example: "/user_photos/luca_dev.jpg"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user:
    get:
      tags:
        - user
      summary: Cerca utenti per username
      description: >-
        Restituisce una lista di utenti il cui username contiene il testo specificato.
      operationId: searchUsers
      parameters:
        - name: search
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 16
            pattern: '^[a-zA-Z0-9_]+$'
          description: Stringa da cercare all'interno degli username (case insensitive)
      responses:
        '200':
          description: Lista utenti corrispondenti al criterio di ricerca
          content:
            application/json:
              schema:
                type: object
                required:
                  - users
                properties:
                  users:
                    type: array
                    minItems: 0
                    maxItems: 200
                    items:
                      type: object
                      properties:
                        username:
                          type: string
                          example: "luca_dev"
                        photoUrl:
                          type: string
                          example: "/user_photos/luca_dev.jpg"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Dati non validi
    Unauthorized:
      description: Accesso negato (token mancante o non valido)
...
    Conflict:
      description: Richiesta in conflitto con lo stato attuale della risposta
    InternalServerError:
      description: Errore interno del server
...
  schemas:
...
    User:
      type: object
      required:
        - uuid
        - username
      description: Identifica un utente
      properties:
        uuid:
          type: string
          format: uuid
          example: 6e9f8a42-1234-5678-90ab-cdef12345678
          description: identificatore univoco utente
        username:
          type: string
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_]+$'
          description: Username dell'utente
        photoUrl:
          type: string
          pattern: '^https?://.*'
          example: 'https://www.example.com/imgs/photo.png'
          description: URL della foto profilo dell’utente
...